day2/trendy_prices/requirements.txt
numpy
scipy
python-kafka
